Drastically changed Items
Items were a pain to make because each item overwrote the
"get<Something>" methods of the base item class. Now that the items
instead return private variables from the item base class, you can
define an item entirely in the constructor. This, of course, destroyed
most items, but that's okay. They were old and not necessarily correct
anyway. See change to, for example, "AncientCoin" for details.