Drastically changed plumbing
Realized the old way of handling heuristics and whatnot totally wasn't
going to work. Much refactoring was done to try and get a more logical
pipeline. "Build" is now the wrapper for most things so that the
"character" can be a flyweight classes. Heuristics are now "Heuristic
Components", and is an enum. Etc.